Moved square_date, square_date_last_four, and split_last_four from Enigma to Offset
