+++ test/enigma_test.rb
@@ -22,20 +22,8 @@ class EnigmaTest < Minitest::Test
   def test_current_date_returns_todays_date_in_MMDDYY
     expected = @today
     assert_equal expected, @enigma.current_date
-  end
-
-  def test_square_date_multiplies_current_date_times_current_date
-    expected = (@today.to_i * @today.to_i).to_s
-    assert_equal expected, @enigma.square_date
-  end
-
-  def test_square_date_last_four_returns_last_four_digits_of_square_date
-    expected = ((@today.to_i * @today.to_i).to_s)[-4..-1]
-    assert_equal expected, @enigma.square_date_last_four
-  end
-
-  def test_split_last_four_returns_array_of_four_offsets
-    assert_equal 4, @enigma.split_last_four.count
+    assert_equal 6, @enigma.current_date.size
+    assert_instance_of String, @enigma.current_date
   end
 
   # def test_key_shifter_adds_split_last_four_and_split_keys
